refactor: resolvedStyleToCss to its own file
